+++ package.py
@@ -1,6 +1,6 @@
 name = "bootycall"
 
-version = "0.20.3"
+version = "0.20.4"
 
 description = "An interface for managing rez environments and resolving packages."